took out sign out button and added functionality to logout button instead
